Updated userinfo, added edit_skills and add_skill
Userinfo displays skills apporpriately for viewers, and provides options
for the skill's owner to edit them.  If another user tries to access the
edit options (which they should not see and can only reach via typing in
the url directly), they are automatically redirected back to home - thus
only the owner should be able to edit their skills.  Minor changes to
the uleague database because 'reference sports_list' was causing errors.
Home is still bugged, and we may need to reert to an earlier version if
not fixed.